+++ lib/dialog-renderer/test.js
@@ -1,6 +1,6 @@
 import { shallow } from 'enzyme';
 
-import DialogRenderer from './dialog-renderer';
+import DialogRenderer from '../dialog-renderer';
 
 describe('DialogRenderer', () => {
   const props = {
